A script to create a SAML certificate
This writes an X.509 certificate outo to `cert-XXX.pem` and a 2048-bit
RSA private key out to `key-XXX.pem`, both in the local directory.
`XXX` is replaced by some random number.  The cert is valid for 729
days.

The certificate is very basic, providing only enough information to get
a SAML IdP to accept it.  The DN is set to bookstack, Stanford University,
Stanford, California, US.  Our srcc-support email is used as the email.
The only extensions are the `CA:false` basic constraint, and the
`serverAuth` extended key usage.

The script also creates a CSR and OpenSSL config file during its work,
but these are deleted.