The script to create LXC container
This script shows (and does) all the things needed to create an LXC
container that is suitable for running our Bookstack containers.

It collects some starting information (hostname, Let's Encrypt info,
Vault AppRole, Git repo), prepares a cloud-init configuration, and
creates the container.

Once the LXC container is created, the user is given the container's MAC
(for entry into NetDB), and told what to do next.