Add a script to fetch from Vault
The `fetch-vault.sh` script uses existing environment variables to fetch
secrets from Vault.  Right now, those secrets are the SAML cert & key,
and the DB root & bookstack passwords.

Environment variables hold all the Vault information we need (address,
KV mount path, base path, AppRole App ID), but not the AppRole Secret
ID.  That might be in the environment, but if not, we prompt for it.

Environment variables are also read to get the paths for the SAML & DB
stuff we fetch from vault.  We overwrite any existing files.  We assume
the base directories are already created.

The Vault token is cleaned up at the end.

The README is also updated to document this script.