Add Restic backups to the stack
Restic is configured to back up to a repository in a Google Cloud Storage
bucket.  That means we have

The LXC creation script now asks for a Google Cloud project ID and a Google
Cloud Storage bucket name.  These can be pre-provided by environment variable.

The Vault fetch script is updated to pull two new secrets (GCP credentials and
Restic repo password) from the Vault.

The Docker Compose file now depends on four new environment variables: GCP
project ID and bucket name, plus paths to the GCP credentials and the Restic
repo password.  Restic is added as a service: It pulls from
[bookstack-restic](https://github.com/stanford-rc/bookstack-restic), runs
`cron`, and depends on the DB container.  It's given full access to the MariaDB
server, and to the Bookstack data volume.

The README is updated to reflect the two new variables used during LXC
creation, and the four new variables used by Docker Compose.  A new entry is
added to the Vault list.  Instructions are
provided for Google Cloud Storage creation (a Bucket and a Service Account),
along with first-time setup instructions for Restic.  Finally, instructions are
provided for upgrading the Restic container.