docker-compose: Set STORAGE_TYPE "local_secure"
The default image storage option puts images in a directory that does
not have any sort of authentication.  That means anyone can donwload an
image, if they have the URL to it.

This changes to the "local_secure" method.  All images are now loaded
through a Bookstack PHP page, which does a login check: The caller has
to be logged in to Bookstack before they can get to the image.

This is not perfect, because the check is only to see if the user is
logged in, not if the user actually has access to a given image.  For
_that_ you need to use the
[`local_secure_restricted`](https://www.bookstackapp.com/docs/admin/upload-config/#local-secure---restricted)
setting, but that's still an experimental setting.

No images have been uploaded yet, so there are no existing images to
[migrate](https://www.bookstackapp.com/docs/admin/upload-config/#migrating-to-secure-images).