docker-compose: Enable removing SAML folks from groups
This turns on `SAML2_REMOVE_FROM_GROUPS`: When a user logs in via SAML,
Bookstack will remove the user from any groups where the user is not a
member of a corresponding SAML group.

This can be problematic to enable, if you have groups that are not
linked to SAML groups (to Workgroups).  That's not a problem for us.