Reconfigure to use SAML2 from the start
This changes the Docker Compose file, and also changes the README, to
use SAML login from the start, when configuring a Bookstack instance
from scratch.

An interesting vulnerability becomes possible when configuring a new
Bookstack instance, which has a Let's Encrypt certificate available from
the start: As part of getting the certificate, our domain name is
published in Certificate Transparency logs.  If the default static
credentials are left enabled, then someone might quickly try to log in
with those credentials.

The workaround is to use SAML authentication from the beginning.